+++ rust/main/agents/relayer/src/msg/processor.rs
@@ -40,6 +40,7 @@ pub struct MessageProcessor {
     destination_ctxs: HashMap<u32, Arc<MessageContext>>,
     metric_app_contexts: Vec<(MatchingList, String)>,
     nonce_iterator: ForwardBackwardIterator,
+    max_retries: u32,
 }
 
 #[derive(Debug)]
@@ -238,6 +239,11 @@ impl Debug for MessageProcessor {
 
 #[async_trait]
 impl ProcessorExt for MessageProcessor {
+    /// The name of this processor
+    fn name(&self) -> String {
+        format!("processor::message::{}", self.domain().name())
+    }
+
     /// The domain this processor is getting messages from.
     fn domain(&self) -> &HyperlaneDomain {
         self.nonce_iterator.high_nonce_iter.db.domain()
@@ -295,12 +301,15 @@ impl ProcessorExt for MessageProcessor {
 
             let app_context = app_context_classifier.get_app_context(&msg).await?;
             // Finally, build the submit arg and dispatch it to the submitter.
-            let pending_msg = PendingMessage::from_persisted_retries(
+            let pending_msg = PendingMessage::maybe_from_persisted_retries(
                 msg,
                 self.destination_ctxs[&destination].clone(),
                 app_context,
+                self.max_retries,
             );
-            self.send_channels[&destination].send(Box::new(pending_msg) as QueueOperation)?;
+            if let Some(pending_msg) = pending_msg {
+                self.send_channels[&destination].send(Box::new(pending_msg) as QueueOperation)?;
+            }
         } else {
             tokio::time::sleep(Duration::from_secs(1)).await;
         }
@@ -319,6 +328,7 @@ impl MessageProcessor {
         send_channels: HashMap<u32, UnboundedSender<QueueOperation>>,
         destination_ctxs: HashMap<u32, Arc<MessageContext>>,
         metric_app_contexts: Vec<(MatchingList, String)>,
+        max_retries: u32,
     ) -> Self {
         Self {
             message_whitelist,
@@ -329,6 +339,7 @@ impl MessageProcessor {
             destination_ctxs,
             metric_app_contexts,
             nonce_iterator: ForwardBackwardIterator::new(Arc::new(db) as Arc<dyn HyperlaneDb>),
+            max_retries,
         }
     }
 
@@ -385,16 +396,16 @@ impl MessageProcessorMetrics {
 mod test {
     use std::time::Instant;
 
-    use crate::{
-        merkle_tree::builder::MerkleTreeBuilder,
-        msg::{
-            gas_payment::GasPaymentEnforcer,
-            metadata::{BaseMetadataBuilder, IsmAwareAppContextClassifier},
+    use prometheus::{IntCounter, Registry};
+    use tokio::{
+        sync::{
+            mpsc::{self, UnboundedReceiver},
+            RwLock,
         },
-        processor::Processor,
+        time::sleep,
     };
+    use tokio_metrics::TaskMonitor;
 
-    use super::*;
     use hyperlane_base::{
         db::{
             test_utils, DbResult, HyperlaneRocksDB, InterchainGasExpenditureData,
@@ -406,16 +417,34 @@ mod test {
         test_utils::dummy_domain, GasPaymentKey, InterchainGasPayment, InterchainGasPaymentMeta,
         MerkleTreeInsertion, PendingOperationStatus, H256,
     };
-    use hyperlane_test::mocks::{MockMailboxContract, MockValidatorAnnounceContract};
-    use prometheus::{IntCounter, Registry};
-    use tokio::{
-        sync::{
-            mpsc::{self, UnboundedReceiver},
-            RwLock,
-        },
-        time::sleep,
+    use hyperlane_operation_verifier::{
+        ApplicationOperationVerifier, ApplicationOperationVerifierReport,
     };
-    use tokio_metrics::TaskMonitor;
+    use hyperlane_test::mocks::{MockMailboxContract, MockValidatorAnnounceContract};
+
+    use crate::{
+        merkle_tree::builder::MerkleTreeBuilder,
+        msg::{
+            gas_payment::GasPaymentEnforcer,
+            metadata::{BaseMetadataBuilder, IsmAwareAppContextClassifier},
+        },
+        processor::Processor,
+    };
+
+    use super::*;
+
+    struct DummyApplicationOperationVerifier {}
+
+    #[async_trait]
+    impl ApplicationOperationVerifier for DummyApplicationOperationVerifier {
+        async fn verify(
+            &self,
+            _app_context: &Option<String>,
+            _message: &HyperlaneMessage,
+        ) -> Option<ApplicationOperationVerifierReport> {
+            None
+        }
+    }
 
     fn dummy_processor_metrics(domain_id: u32) -> MessageProcessorMetrics {
         MessageProcessorMetrics {
@@ -492,11 +521,12 @@ mod test {
         let base_metadata_builder = dummy_metadata_builder(origin_domain, destination_domain, db);
         let message_context = Arc::new(MessageContext {
             destination_mailbox: Arc::new(MockMailboxContract::default()),
-            origin_db: db.clone(),
+            origin_db: Arc::new(db.clone()),
             metadata_builder: Arc::new(base_metadata_builder),
             origin_gas_payment_enforcer: Arc::new(GasPaymentEnforcer::new([], db.clone())),
             transaction_gas_limit: Default::default(),
             metrics: dummy_submission_metrics(),
+            application_operation_verifier: Some(Arc::new(DummyApplicationOperationVerifier {})),
         });
 
         let (send_channel, receive_channel) = mpsc::unbounded_channel::<QueueOperation>();
@@ -510,6 +540,7 @@ mod test {
                 HashMap::from([(destination_domain.id(), send_channel)]),
                 HashMap::from([(destination_domain.id(), message_context)]),
                 vec![],
+                DEFAULT_MAX_MESSAGE_RETRIES,
             ),
             receive_channel,
         )
@@ -783,8 +814,12 @@ mod test {
                 .zip(msg_retries_to_set.iter())
                 .for_each(|(pm, expected_retries)| {
                     // Round up the actual backoff because it was calculated with an `Instant::now()` that was a fraction of a second ago
-                    let expected_backoff = PendingMessage::calculate_msg_backoff(*expected_retries)
-                        .map(|b| b.as_secs_f32().round());
+                    let expected_backoff = PendingMessage::calculate_msg_backoff(
+                        *expected_retries,
+                        DEFAULT_MAX_MESSAGE_RETRIES,
+                        None,
+                    )
+                    .map(|b| b.as_secs_f32().round());
                     let actual_backoff = pm.next_attempt_after().map(|instant| {
                         instant.duration_since(Instant::now()).as_secs_f32().round()
                     });