+++ typescript/cli/src/context/strategies/chain/MultiChainResolver.ts
@@ -82,6 +82,7 @@ export class MultiChainResolver implements ChainResolver {
         warp: argv.warp,
         symbol: argv.symbol,
       });
+
       argv.context.warpCoreConfig = warpCoreConfig;
       const chains = extractChainsFromObj(warpCoreConfig);
       return chains;
@@ -128,32 +129,30 @@ export class MultiChainResolver implements ChainResolver {
     argv: Record<string, any>,
   ): Promise<ChainName[]> {
     const { multiProvider } = argv.context;
-    const chains = [];
+    const chains = new Set<ChainName>();
 
     if (argv.origin) {
-      chains.push(argv.origin);
-    }
-
-    if (argv.destination) {
-      chains.push(argv.destination);
+      chains.add(argv.origin);
     }
 
     if (argv.chain) {
-      chains.push(argv.chain);
+      chains.add(argv.chain);
     }
 
-    if (!argv.chains && chains.length === 0) {
+    if (argv.chains) {
+      const additionalChains = argv.chains
+        .split(',')
+        .map((item: string) => item.trim());
+      return Array.from(new Set([...chains, ...additionalChains]));
+    }
+
+    // If no destination is specified, return all EVM chains
+    if (!argv.destination) {
       return Array.from(this.getEvmChains(multiProvider));
     }
 
-    return Array.from(
-      new Set([
-        ...chains,
-        ...(argv.chains
-          ? argv.chains.split(',').map((item: string) => item.trim())
-          : []),
-      ]),
-    );
+    chains.add(argv.destination);
+    return Array.from(chains);
   }
 
   private async getWarpRouteConfigChains(